+++ packages/tswebextension/src/lib/mv2/background/services/cookie-filtering/parsed-cookie.ts
@@ -1,4 +1,4 @@
-import { parse } from 'tldts';
+import { type Cookies } from 'webextension-polyfill';
 
 /**
  * Synthetic Cookie-like object parsed from headers.
@@ -10,9 +10,16 @@ export default class ParsedCookie {
     url: string;
 
     /**
-     * The domain of the cookie.
+     * Defines the host to which the cookie will be sent.
+     * Only the current domain can be set as the value, or a domain of a higher
+     * order, unless it is a public suffix. Setting the domain will make
+     * the cookie available to it, as well as to all its subdomains.
+     *
+     * Can be empty. And actually should be empty for `__Host-` cookies.
+     *
+     * @see {@link https://developer.mozilla.org/en-US/docs/Web/HTTP/Headers/Set-Cookie#attributes}
      */
-    domain: string;
+    domain?: string;
 
     /**
      * The name of the cookie.
@@ -30,7 +37,7 @@ export default class ParsedCookie {
     expires?: Date;
 
     /**
-     * The parsed max-age value.
+     * The parsed max-age value in seconds.
      */
     maxAge?: number;
 
@@ -50,7 +57,7 @@ export default class ParsedCookie {
     sameSite?: string;
 
     /**
-     * Cookies path.
+     * Cookies path. Defaults to the path portion of the url parameter.
      */
     path?: string;
 
@@ -78,7 +85,38 @@ export default class ParsedCookie {
         this.value = value;
 
         this.url = url;
-        const tldResult = parse(url);
-        this.domain = tldResult.domain!;
+        this.path = new URL(url).pathname;
+    }
+
+    /**
+     * Creates new {@link ParsedCookie} from provided {@link Cookies.Cookie}.
+     *
+     * @param cookie Item of {@link Cookies.Cookie}.
+     * @param url String URL relative to this cookie.
+     *
+     * @returns New {@link ParsedCookie}.
+     */
+    static fromBrowserCookie(cookie: Cookies.Cookie, url: string): ParsedCookie {
+        const parsedCookie = new ParsedCookie(cookie.name, cookie.value, url);
+
+        parsedCookie.thirdParty = !cookie.firstPartyDomain;
+        parsedCookie.httpOnly = cookie.httpOnly;
+        parsedCookie.secure = cookie.secure;
+        parsedCookie.sameSite = cookie.sameSite;
+
+        // For hostOnly cookie domain should be empty and path should be '/'.
+        if (cookie.hostOnly) {
+            parsedCookie.path = '/';
+        } else {
+            parsedCookie.domain = cookie.domain;
+            parsedCookie.path = cookie.path;
+        }
+
+        if (cookie.expirationDate) {
+            // cookie.expirationDate is a number of seconds that's why it should be multiplied
+            parsedCookie.expires = new Date(cookie.expirationDate * 1000);
+        }
+
+        return parsedCookie;
     }
 }